+++ dev/e2e_app/patrol_test/camera_features/take_a_photo_test.dart
@@ -7,13 +7,13 @@ void main() {
   patrol('take a photo', ($) async {
     await createApp($);
     final cameraHelpers = CameraHelpers($);
+
     await $(#cameraFeaturesButton).scrollTo().tap();
-    if (await $.platform.mobile.isVirtualDevice() && Platform.isIOS) {
-      throw Exception('Camera is not supported on iOS simulator');
-    }
     await $(#takePhotoButton).tap();
     await cameraHelpers.maybeAcceptPermissionDialog();
-    await cameraHelpers.maybeAcceptDialogAndroid();
+    if (Platform.isAndroid) {
+      await cameraHelpers.maybeAcceptDialogAndroid();
+    }
 
     await $.platform.mobile.takeCameraPhoto();
     await $.pumpAndSettle();