+++ packages/patrol/lib/src/platform/patrol_app_service.dart
@@ -1,2 +1,2 @@
 export 'mobile/patrol_app_service_io.dart'
-    if (dart.library.html) 'web/patrol_app_service_web.dart';
+    if (dart.library.js_interop) 'web/patrol_app_service_web.dart';