+++ packages/patrol/CHANGELOG.md
@@ -1,8 +1,8 @@
 ## Unreleased
 
-- Reflect failed tests in Playwright report. (#2970)
 - Fix `tapOnNotification` on iOS physical devices to Open button when tapping on notification. (#2972)
 - Fix `takeCameraPhoto` do not accept location permission dialog on Pixel physical devices. (#2972)
+- Reflect failed tests in Playwright report. (#2970)
 
 ## 4.3.0